scheduler/pyproject.toml
[project]
name = "scheduler"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=0.3.12",
    "langchain-text-splitters>=0.3.8",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
]
